+++ features/ftp_server/step_definitions/test_server.rb
@@ -22,3 +22,12 @@ Given /^the test server has auth level "(.*?)"$/ do |auth_level|
   auth_level = Ftpd.const_get(auth_level)
   server.auth_level = auth_level
 end
+
+Given /^the test server has session timeout set to (\S+) seconds$/ do
+|timeout|
+  server.session_timeout = timeout.to_f
+end
+
+Given /^the test server has session timeout disabled$/ do
+  server.session_timeout = nil
+end